+++ time/src/main/java/net/jqwik/time/providers/DatesArbitraryProvider.java
@@ -1,4 +1,4 @@
-package net.jqwik.time;
+package net.jqwik.time.providers;
 
 import java.time.*;
 import java.util.*;
@@ -7,14 +7,14 @@ import net.jqwik.api.*;
 import net.jqwik.api.providers.*;
 import net.jqwik.time.api.*;
 
-public class YearMonthArbitraryProvider implements ArbitraryProvider {
+public class DatesArbitraryProvider implements ArbitraryProvider {
 	@Override
 	public boolean canProvideFor(TypeUsage targetType) {
-		return targetType.isAssignableFrom(YearMonth.class);
+		return targetType.isAssignableFrom(LocalDate.class);
 	}
 
 	@Override
 	public Set<Arbitrary<?>> provideFor(TypeUsage targetType, SubtypeProvider subtypeProvider) {
-		return Collections.singleton(Dates.yearMonths());
+		return Collections.singleton(Dates.dates());
 	}
 }
