+++ time/src/main/java/net/jqwik/time/providers/YearMonthArbitraryProvider.java
@@ -1,4 +1,4 @@
-package net.jqwik.time;
+package net.jqwik.time.providers;
 
 import java.time.*;
 import java.util.*;
@@ -7,14 +7,14 @@ import net.jqwik.api.*;
 import net.jqwik.api.providers.*;
 import net.jqwik.time.api.*;
 
-public class YearArbitraryProvider implements ArbitraryProvider {
+public class YearMonthArbitraryProvider implements ArbitraryProvider {
 	@Override
 	public boolean canProvideFor(TypeUsage targetType) {
-		return targetType.isAssignableFrom(Year.class);
+		return targetType.isAssignableFrom(YearMonth.class);
 	}
 
 	@Override
 	public Set<Arbitrary<?>> provideFor(TypeUsage targetType, SubtypeProvider subtypeProvider) {
-		return Collections.singleton(Dates.years());
+		return Collections.singleton(Dates.yearMonths());
 	}
 }
